Conflicts:
	src/Battlescape/ActionMenuState.cpp
	src/Battlescape/InventoryState.cpp
	src/Battlescape/Map.cpp
	src/Battlescape/Projectile.cpp
	src/Battlescape/Projectile.h
	src/Battlescape/ProjectileFlyBState.cpp
	src/Battlescape/ProjectileFlyBState.h
	src/Battlescape/UnitInfoState.cpp
	src/Geoscape/GraphsState.cpp
	src/Geoscape/MonthlyReportState.cpp
	src/Interface/ArrowButton.h
	src/Interface/ToggleTextButton.cpp
	src/Interface/ToggleTextButton.h
	src/Menu/DeleteGameState.cpp
	src/Menu/DeleteGameState.h
	src/Menu/LoadState.cpp
	src/Menu/SaveState.cpp
	src/Menu/SaveState.h
	src/Menu/SavedGameState.cpp
	src/Menu/SavedGameState.h
	src/Ruleset/Ruleset.h
	src/Savegame/BattleUnit.cpp
	src/Savegame/BattleUnit.h
	src/Savegame/SavedGame.cpp
	src/Savegame/SavedGame.h

+++ src/Savegame/BattleUnit.h
@@ -389,11 +389,11 @@ private:
 		void clearVisibleTiles();
 
 		/// Calculate firing accuracy.
-		double getFiringAccuracy(
+		int getFiringAccuracy(
 				BattleActionType actionType,
 				BattleItem* item);
 		/// Calculate accuracy modifier.
-		double getAccuracyModifier();
+		int getAccuracyModifier();
 		/// Calculate throwing accuracy.
 		double getThrowingAccuracy();
